chore: clear codes and rename plugins
