+++ src-plugins/cad-plugin/README.md
