+++ src-plugins/cad-plugin/package.json
