+++ src-plugins/cad-plugin/src/i18n/en.ts
