+++ src-plugins/cad-plugin/src/i18n/index.ts
