+++ src-plugins/cad-plugin/src/i18n/zh.ts
