+++ src-plugins/cad-plugin/src/plugin-types.ts
