+++ src-plugins/cad-plugin/vite.config.ts
